🏗 test-status app refactors + report e2e tests (#22501)
* Report e2e test runs to the test-status app

* Move `status-report.js` out of `runtime-test`

* Address feedback

* Fix path issues

* Address feedback
