+++ build-system/tasks/mocha-ci-reporter.js
@@ -17,6 +17,7 @@
 
 const {Base} = require('mocha').reporters;
 const {inherits} = require('mocha').utils;
+const {reportTestFinished} = require('./report-test-status');
 const {symbols} = require('./karma.conf').mochaReporter;
 
 /**
@@ -46,6 +47,8 @@ function ciReporter(runner) {
 
   function epilogue() {
     const {failures, stats} = self;
+    reportTestFinished(stats.passes, stats.failures);
+
     Base.list(failures);
     process.stdout.write(
       `Executed ${stats.failures + stats.passes} of ${stats.tests} ` +
